VorbisComments: Error traces
