+++ lofty/src/flac/mod.rs
@@ -14,8 +14,8 @@ use crate::config::WriteOptions;
 use crate::error::{LoftyError, Result};
 use crate::file::{FileType, TaggedFile};
 use crate::id3::v2::tag::Id3v2Tag;
-use crate::ogg::tag::VorbisCommentsRef;
-use crate::ogg::{OggPictureStorage, VorbisComments};
+use crate::ogg::OggPictureStorage;
+use crate::ogg::tag::{VorbisComments, VorbisCommentsRef};
 use crate::picture::{Picture, PictureInformation};
 use crate::tag::TagExt;
 use crate::util::io::{FileLike, Length, Truncate};
