+++ lofty/tests/taglib/test_flac.rs
@@ -7,7 +7,8 @@ use lofty::config::{ParseOptions, WriteOptions};
 use lofty::file::AudioFile;
 use lofty::flac::FlacFile;
 use lofty::id3::v2::Id3v2Tag;
-use lofty::ogg::{OggPictureStorage, VorbisComments};
+use lofty::ogg::OggPictureStorage;
+use lofty::ogg::tag::VorbisComments;
 use lofty::picture::{MimeType, Picture, PictureInformation, PictureType};
 use lofty::tag::{Accessor, TagExt};
